+++ src/plugin/index.ts
@@ -2,3 +2,4 @@ export * from "./factory/VideoID";
 export * from "./factory/url";
 export * from "./factory/toTop";
 export * from "./factory/auto-update";
+export * from "./factory/stats-generator";