+++ src/tsTypes/index.d.ts
@@ -1,1 +1,5 @@
 export * from "./links/CreateLinks";
+
+// blog
+export * from "./blog/BlogItem";
+export * from "./blog/ContentCache";